Test de integracion uh2
happy y sad path

ademas se refactorizo los test unitarios para que esten en la misma clase
