+++ src/test/java/com/bootcamp/be_java_hisp_w29_g3/service/SocialMeliServiceImplTest.java
@@ -1,7 +1,10 @@
 package com.bootcamp.be_java_hisp_w29_g3.service;
 
 import com.bootcamp.be_java_hisp_w29_g3.dto.response.FollowDto;
+import com.bootcamp.be_java_hisp_w29_g3.dto.response.FollowerCountDTO;
 import com.bootcamp.be_java_hisp_w29_g3.dto.response.UnfollowDto;
+import com.bootcamp.be_java_hisp_w29_g3.entity.Buyer;
+import com.bootcamp.be_java_hisp_w29_g3.entity.Seller;
 import com.bootcamp.be_java_hisp_w29_g3.exception.BadRequestException;
 import com.bootcamp.be_java_hisp_w29_g3.exception.NotFoundException;
 import com.bootcamp.be_java_hisp_w29_g3.repository.IUserRepository;
@@ -11,6 +14,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -138,6 +144,64 @@ class SocialMeliServiceImplTest {
 
         //Assert + Act
         assertThrows(BadRequestException.class,()-> socialMeliService.unfollowSeller(userId,userIdToUnFollow));
+    }
+
+    @DisplayName("US-0002 - Happy path with many followers (T-0007)")
+    @Test
+    void calculateSellerFollowerCount(){
+        //arrange
+        Integer sellerId = 1;
+        String sellerName = "Juan";
+        Seller seller = Seller.builder().id(sellerId).name(sellerName).build();
+
+        List<Buyer> buyerList = List.of(Buyer.builder().build(),Buyer.builder().build());
+
+        FollowerCountDTO expected = new FollowerCountDTO(sellerId,sellerName, (long) buyerList.size());
+
+        when(userRepository.existsSellerById(sellerId)).thenReturn(true);
+        when(userRepository.getSellerById(sellerId)).thenReturn(seller);
+        when(userRepository.getBuyersFollowingSeller(sellerId)).thenReturn(buyerList);
+
+        //act
+
+        FollowerCountDTO obtained = socialMeliService.calculateSellerFollowerCount(sellerId);
+
+        //assert
+        assertEquals(expected,obtained);
+    }
+    @DisplayName("US-0002 - Seller not found (T-0007)")
+    @Test
+    void calculateSellerFollowerCountNotFound(){
+        //arrange
+        Integer sellerId = 1;
+
+        when(userRepository.existsSellerById(sellerId)).thenReturn(false);
+
+        //act & assert
+        assertThrows(NotFoundException.class, () -> socialMeliService.calculateSellerFollowerCount(sellerId));
+    }
+
+    @DisplayName("US-0002 - Happy path with many followers (T-0007)")
+    @Test
+    void calculateSellerFollowerCountEqual0(){
+        //arrange
+        Integer sellerId = 1;
+        String sellerName = "Juan";
+        Seller seller = Seller.builder().id(sellerId).name(sellerName).build();
+
+        List<Buyer> buyerList = new ArrayList<>();
+
+        FollowerCountDTO expected = new FollowerCountDTO(sellerId,sellerName, (long) buyerList.size());
+
+        when(userRepository.existsSellerById(sellerId)).thenReturn(true);
+        when(userRepository.getSellerById(sellerId)).thenReturn(seller);
+        when(userRepository.getBuyersFollowingSeller(sellerId)).thenReturn(buyerList);
+
+        //act
+        FollowerCountDTO obtained = socialMeliService.calculateSellerFollowerCount(sellerId);
+
+        //assert
+        assertEquals(expected,obtained);
     }
